[refactor] header의 모달에 넣어줄 header, body, footer 모달 컴포넌트 분리
