+++ client/src/components/common/Header/UserProfileModalBox/index.ts
@@ -1,0 +1,3 @@
+export { default as UserProfileModalHeader } from './UserProfileHeader/UserProfileHeader';
+export { default as UserProfileModalBody } from './UserProfileBody/UserProfileBody';
+export { default as UserProfileModalFooter } from './UserProfileFooter/UserProfileFooter';
